Færði yfir í ChatGPT Assistant fyrir undirbúningshluta
