+++ .idea/vcs.xml
@@ -2,6 +2,5 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/src/main/java/vv/intertrain/vinnsla/assistant_api_java" vcs="Git" />
   </component>
 </project>
